# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AlgoJudge.Server",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/identity/providers/{slug}/challenge": {
      "get": {
        "tags": [
          "FederatedSignIn"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/identity/providers/{slug}/signed-in": {
      "get": {
        "tags": [
          "FederatedSignIn"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "returnUrl",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect"
          }
        }
      }
    },
...
    "/identity/providers/{id}": {
      "get": {
        "tags": [
          "IdentityProviders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProviderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProviderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProviderDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "IdentityProviders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityProviderInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityProviderInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityProviderInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProviderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProviderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProviderDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "IdentityProviders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IdentityProviderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "issuer": {
            "type": "string",
            "nullable": true
          },
          "clientId": {
            "type": "string",
            "nullable": true
          },
          "scopes": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "accountUrl": {
            "type": "string",
            "nullable": true
          },
          "claimPath": {
            "type": "string",
            "nullable": true
          },
          "unmappedBehavior": {
            "type": "string",
            "nullable": true
          },
          "defaultTemplateName": {
            "type": "string",
            "nullable": true
          },
          "deletionChannelEnabled": {
            "type": "boolean"
          },
          "hasClientSecret": {
            "type": "boolean"
          },
          "hasDeletionSecret": {
            "type": "boolean"
          },
          "callbackPath": {
            "type": "string",
            "nullable": true
          },
          "mappingRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MappingRuleDto"
            },
            "nullable": true
          },
          "linkedAccounts": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IdentityProviderInputDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "issuer": {
            "type": "string",
            "nullable": true
          },
          "clientId": {
            "type": "string",
            "nullable": true
          },
          "clientSecret": {
            "type": "string",
            "nullable": true
          },
          "scopes": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "accountUrl": {
            "type": "string",
            "nullable": true
          },
          "claimPath": {
            "type": "string",
            "nullable": true
          },
          "unmappedBehavior": {
            "type": "string",
            "nullable": true
          },
          "defaultTemplateName": {
            "type": "string",
            "nullable": true
          },
          "deletionChannelEnabled": {
            "type": "boolean"
          },
          "deletionSecret": {
            "type": "string",
            "nullable": true
          },
          "mappingRules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MappingRuleDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MappingRuleDto": {
        "type": "object",
        "properties": {
          "claimValue": {
            "type": "string",
            "nullable": true
          },
          "templateName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}